Prepare dicom and meta
