+++ scripts/prepare_meta_dicom.py
@@ -10,16 +10,14 @@ import cv2
 import pydicom
 from tqdm import tqdm
 from joblib import delayed, Parallel
+import zipfile
+from pydicom.filebase import DicomBytesIO
+from scripts.logs import get_logger
+from scripts.utils import dumpobj, loadobj, GradualWarmupScheduler
 
-
-def dumpobj(file, obj):
-    with open(file, 'wb') as handle:
-        pickle.dump(obj, handle, protocol=pickle.HIGHEST_PROTOCOL)
-        
-def loadobj(file):
-    with open(file, 'rb') as handle:
-        return pickle.load(handle)
-
+# Print info about environments
+logger = get_logger('Prepare Data', 'INFO') # noqa
+logger.info('Cuda set up : time {}'.format(datetime.datetime.now().time()))
 
 def get_dicom_value(x, cast=int):
     if type(x) in [pydicom.multival.MultiValue, tuple]:
@@ -79,50 +77,6 @@ def apply_window_policy(image):
 
     return image
 
-
-def apply_dataset_policy(df, policy):
-    if policy == 'all':
-        pass
-    elif policy == 'pos==neg':
-        df_positive = df[df.labels != '']
-        df_negative = df[df.labels == '']
-        df_sampled = df_negative.sample(len(df_positive))
-        df = pd.concat([df_positive, df_sampled], sort=False)
-    else:
-        raise
-    log('applied dataset_policy %s (%d records)' % (policy, len(df)))
-
-    return df
-
-def autocrop(image, threshold=0):
-    """Crops any edges below or equal to threshold
-    Crops blank image to 1x1.
-    Returns cropped image.
-    https://stackoverflow.com/questions/13538748/crop-black-edges-with-opencv
-    """
-
-    flatImage = np.max(image, 2)
-    rows = np.where(np.max(flatImage, 0) > flatImage.min())[0]
-    cols = np.where(np.max(flatImage, 1) > flatImage.min())[0]
-    image = image[cols[0]: cols[-1] + 1, rows[0]: rows[-1] + 1]
-    sqside = max(image.shape)
-    minchl = np.min(image, axis=(0,1))
-    imageout = np.ones((sqside, sqside, 3)) * minchl
-    imageout[:image.shape[0], :image.shape[1],:] = image.copy()
-    return imageout
-
-def convert_dicom_to_npz(imfile):
-    try:
-        imgnm = (imfile.split('/')[-1]).replace('.dcm', '')
-        dicom = pydicom.dcmread(os.path.join(imfile))
-        image = dicom.pixel_array
-        image = rescale_image(image, rescaledict['RescaleSlope'][imgnm], rescaledict['RescaleIntercept'][imgnm])
-        image = apply_window_policy(image)
-        np.savez_compressed(os.path.join(path_proc, imgnm), image)
-    except:
-        print(imfile)
-
-        
 def convert_dicom_to_jpg(name):
     try:
         data = f.read(name)
@@ -134,27 +88,55 @@ def convert_dicom_to_jpg(name):
         image = apply_window_policy(image)
         image -= image.min((0,1))
         image = (255*image).astype(np.uint8)
-        cv2.imwrite(os.path.join(path_proc, dirtype, imgnm)+'.jpg', image)
+        cv2.imwrite(os.path.join(PATHPROC, dirtype, imgnm)+'.jpg', image)
     except:
-        print(name)
+        logger.info(name)
+        
+def generate_df(base, files):
+    train_di = {}
 
-path_img = '/Users/dhanley2/Documents/Personal/rsna/data/orig'
-BASE_PATH = path_img = '/root/data'
-TRAIN_DIR=os.path.join(path_img, 'stage_1_train_images')
-TEST_DIR=os.path.join(path_img, 'stage_1_test_images')
-path_data = '/Users/dhanley2/Documents/Personal/rsna/data'
-path_data = '/root/data/rsna/data'
-path_proc = '/Users/dhanley2/Documents/Personal/rsna/data/proc'
-path_proc = '/root/data/proc'
+    for filename in tqdm(files):
+        path = os.path.join( base ,  filename)
+        dcm = pydicom.dcmread(path)
+        all_keywords = dcm.dir()
+        ignored = ['Rows', 'Columns', 'PixelData']
 
-trnmdf = pd.read_csv(os.path.join(path_data, 'train_metadata.csv'))
-tstmdf = pd.read_csv(os.path.join(path_data, 'test_metadata.csv'))
+        for name in all_keywords:
+            if name in ignored:
+                continue
+
+            if name not in train_di:
+                train_di[name] = []
+
+            train_di[name].append(dcm[name].value)
+
+    df = pd.DataFrame(train_di)
+    
+    return df
+
+DATAPATH = 'data'
+TRAIN_DIR = os.path.join(DATAPATH, 'raw/stage_1_train_images')
+TEST_DIR = os.path.join(DATAPATH, 'raw/stage_2_test_images')
+PATHPROC = os.path.join(DATAPATH, 'proc')
+
+logger.info('Create test meta files')
+test_files = os.listdir(TEST_DIR)
+test_df = generate_df(TEST_DIR, test_files)
+test_df.to_csv(os.path.join(DATAPATH, 'test_metadata.csv'))
+
+logger.info('Create train meta files')
+train_files = os.listdir( TRAIN_DIR)
+train_df = generate_df(TRAIN_DIR, train_files)
+train_df.to_csv(os.path.join(DATAPATH, 'train_metadata.csv'))
+
+logger.info('Load meta files')
+trnmdf = pd.read_csv(os.path.join(DATAPATH, 'train_metadata.csv'))
+tstmdf = pd.read_csv(os.path.join(DATAPATH, 'test_metadata.csv'))
 mdf = pd.concat([trnmdf, tstmdf], 0)
 rescaledict = mdf.set_index('SOPInstanceUID')[['RescaleSlope', 'RescaleIntercept']].to_dict()
 
-import zipfile
-from pydicom.filebase import DicomBytesIO
-with zipfile.ZipFile(os.path.join(path_img, "rsna-intracranial-hemorrhage-detection.zip"), "r") as f:
+logger.info('Create windowed images')
+with zipfile.ZipFile(os.path.join(DATAPATH, "raw/rsna-intracranial-hemorrhage-detection.zip"), "r") as f:
     for t, name in enumerate(tqdm(f.namelist())):
         convert_dicom_to_jpg(name)
